Fix #11
The app is not automatically added when it is updated but only with
the first installation.
Before, if the user removed an app, it was re-added with the update,
but now if the user has decided to remove it, he/she must put it
back manually.
The previous behavior has not been altered, only improved.